requirements.txt
python-socketio[client]>=5.11.4
